// types.ts
export interface ExamConfiguration {
  moduleId: string;
  moduleTitle: string;
  durationMinutes?: number; 
  timeLimitEnabled: boolean;
  numMultipleChoice: number;
  // numShortAnswer: number; // Removed
  difficulty: number; // 1-5
  initialConfig?: Partial<ExamConfiguration>; 
}